rename perf_counter ot stats
