+++ src/ray/metrics/metric_defs.h
@@ -1,8 +1,6 @@
 // Use `RAY_METRIC(name, description, unit)` to define a metric
 // which parameters are `metric_name`, `description` and `unit`.
 
-// static Metric RedisLatency("redis_latency", "the latency of redis", "ms");
-
 static Metric TaskElapse = Metric::MakeHistogram("task_elapse",
                                                  "the task elapse in raylet",
                                                  "ms",
@@ -15,15 +13,6 @@ static Metric RedisLatency = Metric::MakeHistogram("redis_latency",
                                                    {0, 25, 50, 75, 100, 200},
                                                    {JobNameKey, NodeAddressKey});
 
-//static Metric TaskElapse("task_elapse", "the task elapse in raylet", "ms");
-
-/*
-static Metric TaskElapse = Metric::MakeHistogram("task_elapse",
-                                                 "the task elapse in raylet",
-                                                 "ms",
-                                                 {0, 25, 50, 75, 100, 200},
-                                                 {JobNameKey, NodeAddressKey});
-*/
 
 static Metric TaskCount = Metric::MakeGauge("task_count",
                                             "the task count that the raylet received",
